requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: linux-64
_libgcc_mutex=0.1=conda_forge
_openmp_mutex=4.5=1_llvm
absl-py=1.0.0=pypi_0
addict=2.2.1=pypi_0
ca-certificates=2022.2.1=h06a4308_0
cachetools=4.2.4=pypi_0
certifi=2021.5.30=py36h06a4308_0
cffi=1.14.6=py36hc120d54_0
charset-normalizer=2.0.12=pypi_0
cudatoolkit=11.5.1=hcf5317a_10
cudnn=8.2.1.32=h86fa8c9_0
cycler=0.11.0=pyhd3eb1b0_0
cython=0.29.4=pypi_0
dataclasses=0.8=pyh787bdff_2
dbus=1.13.18=hb2f20db_0
decorator=4.4.2=pypi_0
expat=2.4.4=h295c915_0
fontconfig=2.13.1=h6c09931_0
freetype=2.11.0=h70c0345_0
future=0.18.2=py36h5fab9bb_3
glib=2.69.1=h4ff587b_1
google-auth=2.6.2=pypi_0
google-auth-oauthlib=0.4.6=pypi_0
grpcio=1.44.0=pypi_0
gst-plugins-base=1.14.0=h8213a91_2
gstreamer=1.14.0=h28cd5cc_2
icu=58.2=he6710b0_3
idna=3.3=pypi_0
imagecorruptions=1.1.2=pypi_0
imageio=2.15.0=pypi_0
imgaug=0.4.0=pypi_0
importlib-metadata=4.8.3=pypi_0
jpeg=9e=h7f98852_0
kiwisolver=1.3.1=py36h2531618_0
lcms2=2.12=hddcbb42_0
ld_impl_linux-64=2.35.1=h7274673_9
libblas=3.9.0=12_linux64_mkl
libcblas=3.9.0=12_linux64_mkl
libffi=3.3=he6710b0_2
libgcc-ng=11.2.0=h1d223b6_14
liblapack=3.9.0=12_linux64_mkl
libpng=1.6.37=h21135ba_2
libprotobuf=3.18.0=h780b84a_1
libstdcxx-ng=11.2.0=he4da1e4_14
libtiff=4.2.0=hf544144_3
libuuid=1.0.3=h7f8727e_2
libwebp-base=1.2.2=h7f98852_1
libxcb=1.14=h7b6447c_0
libxml2=2.9.12=h03d6c58_0
llvm-openmp=12.0.1=h4bd325d_1
lz4-c=1.9.3=h9c3ff4c_1
magma=2.5.4=h6103c52_2
markdown=3.3.6=pypi_0
matplotlib=3.3.4=py36h06a4308_0
matplotlib-base=3.3.4=py36h62a2d02_0
mkl=2021.4.0=h8d4b97c_729
mmcv=1.4.7=pypi_0
mmdet=0.5.6+ecfb7d4=dev_0
nccl=2.12.7.1=h0800d71_0
ncurses=6.3=h7f8727e_2
networkx=2.5.1=pypi_0
ninja=1.10.2=h4bd325d_1
numpy=1.19.2=pypi_0
oauthlib=3.2.0=pypi_0
olefile=0.46=pyh9f0ad1d_1
opencv-python=4.5.5.64=pypi_0
openjpeg=2.4.0=hb52868f_1
openssl=1.1.1m=h7f8727e_0
pcre=8.45=h295c915_0
pillow=8.4.0=pypi_0
pip=21.2.2=py36h06a4308_0
protobuf=3.19.4=pypi_0
pyasn1=0.4.8=pypi_0
pyasn1-modules=0.2.8=pypi_0
pycocotools=2.1=pypi_0
pycparser=2.21=pyhd8ed1ab_0
pyparsing=3.0.4=pyhd3eb1b0_0
pyqt=5.9.2=py36h05f1152_2
python=3.6.13=h12debd9_1
python-dateutil=2.8.2=pyhd3eb1b0_0
python_abi=3.6=2_cp36m
pytorch=1.9.1=cuda112py36h36e649e_3
pytorch-gpu=1.9.1=cuda112py36h0bbbad9_3
pywavelets=1.1.1=pypi_0
pyyaml=5.1.2=pypi_0
qt=5.9.7=h5867ecd_1
readline=8.1.2=h7f8727e_1
requests=2.27.1=pypi_0
requests-oauthlib=1.3.1=pypi_0
rsa=4.8=pypi_0
scikit-image=0.17.2=pypi_0
scipy=1.5.4=pypi_0
setuptools=58.0.4=py36h06a4308_0
shapely=1.8.5.post1=pypi_0
sip=4.19.8=py36hf484d3e_0
six=1.16.0=pyhd3eb1b0_1
sleef=3.5.1=h9b69904_2
sqlite=3.38.0=hc218d9a_0
tbb=2021.5.0=h4bd325d_0
tensorboard=2.8.0=pypi_0
tensorboard-data-server=0.6.1=pypi_0
tensorboard-plugin-wit=1.8.1=pypi_0
terminaltables=3.1.0=pypi_0
tifffile=2020.9.3=pypi_0
timm=0.5.4=pypi_0
tk=8.6.11=h1ccaba5_0
torchvision=0.10.1=py36cuda112h71db1b3_0_cuda
tornado=6.1=py36h27cfd23_0
typing_extensions=4.1.1=pyha770c72_0
urllib3=1.26.8=pypi_0
werkzeug=2.0.3=pypi_0
wheel=0.37.1=pyhd3eb1b0_0
xz=5.2.5=h7b6447c_0
zipp=3.6.0=pypi_0
zlib=1.2.11=h7f8727e_4
zstd=1.5.0=ha95c52a_0
